refactor: move manage project files to be consistent with other pages
